requirements.txt
attrs==23.2.0
backcall==0.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
executing==2.0.1
fastjsonschema==2.20.0
idna==3.7
Jinja2==3.1.4
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
pure_eval==0.2.3
Pygments==2.18.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
python-dateutil==2.9.0.post0
pywin32==306
pyzmq==26.0.3
referencing==0.35.1
requests==2.32.3
rpds-py==0.19.0
six==1.16.0
soupsieve==2.5
tinycss2==1.3.0
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.2
wcwidth==0.2.13
webencodings==0.5.1
zipp==3.19.2
